amts-backend/requirements.txt
fastapi==0.128.0
uvicorn==0.39.0
sqlalchemy==2.0.45
alembic==1.16.5
python-dotenv==1.2.1
httpx==0.28.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
python-multipart==0.0.20
pydantic==2.12.5
pydantic-settings==2.11.0
supabase==2.27.0
openai==2.14.0
stripe==14.1.0
aiofiles==25.1.0

